DefaultEventAdapter should be renamed to ExampleEventAdapter and moved to music.py
